python/pyproject.toml
[project]
name = "lang-image-benchmark"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "numpy",
    "opencv-python",
    "Pillow",
    "numba",
    "scipy",
]
